LayoutTests/imported/w3c/web-platform-tests/html/rendering/the-details-element/details-pseudo-elements-006-expected.txt

summary
contents

FAIL ::details-content matches :hover when mouse pointer is over it assert_equals: :hover styles when pointer is inside expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
